requirements.txt
langchain
pypdf
boto3
awscli
chromadb
langchain-chroma # integração com embeddings do bedrock
langchain-community # integração com embeddings do bedrock
python-dotenv
streamlit # teste
langchain-aws
python-telegram-bot
httpx